Replace v1 with v2 of the data and importers
v2 onwards of the data is machine generated and we expect little to no
change to our database schema going forwards.  The delta between v1 and
v2 schemas was significant enough that the v1 importers can no longer
work so it doesn't make sense to keep them around.
